% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_first_text.R
\name{get_first}
\alias{get_first}
\title{Extract texts from one "first" scraped html file to check and set paramters for the get_text function.}
\usage{
get_first(
  loc_html,
  x_text,
  x_title,
  x_date,
  speaker,
  country,
  country_id,
  regime,
  source
)
}
\arguments{
\item{loc_html}{A character vector which designates the location of the html files to extract, e.g. in a folder called "aze_speech/"}

\item{x_text}{The xpath of the text (to be found out, for example, by using the "inspect" tool of firefox)}

\item{x_title}{The xpath of the title of the text
(to be found out, for example, by using the "inspect" tool of firefox)}

\item{x_date}{The xpath of the text's date, if available
(to be found out, for example, by using the "inspect" tool of firefox).
Please NOTE: One might like to use regular expressions and/or the as.Date function
to further adjust the format of the date once extracted.}

\item{speaker}{The name of the speaker (assuming official speeches are extracted from htmls)}

\item{country}{The name of the country of the speaker
(assuming official speeches are extracted from htmls)}

\item{country_id}{A short country identifier,
useful for merging the speech corpus with other datasets during the analysis.}

\item{regime}{The regime type of the speaker's country
(assuming official speeches are extracted from htmls), e.g. autocracy.}

\item{source}{The webpage from which the htmls were scraped
(crucial to always indicate due to copyright reasons!)}
}
\value{
A quanteda corpus consisting of one single "first" text scraped as corpus documents
(for quanteda tutorials on text analysis see https://quanteda.io/index.html)
}
\description{
A helper function to extract one "first" text from an html file as preparation and check of using the function get_text
(finding the right xpath can be tricky sometimes...)
}
\examples{
#Don't run
#Extract texts from htmls
#get_first()

}
