% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links_content.R
\name{get_pagesource}
\alias{get_pagesource}
\title{Automatically collect the page sources of links.}
\usage{
get_pagesource(all_links, case, time_out = 1)
}
\arguments{
\item{all_links}{A character vector which contains all links with the desired content to be downloaded}

\item{case}{The name of the case, e.g. a country, make it a unique case identifier.}

\item{time_out}{By default 1 sec, simulates human action and prevents overloading servers (too many requests in too little time)}
}
\value{
A character vector with all page sources.
}
\description{
A function to automatically collect the page sources of links (typically done before the content of the links is downloaded).
}
\examples{
#Don't run
#Get the page sources of links provided on a website
#get_pagesources()

}
