% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_first_text.R
\name{get_text}
\alias{get_text}
\title{Extract texts from scraped html files.}
\usage{
get_text(
  loc_html,
  x_text,
  x_title,
  x_date,
  speaker,
  country,
  country_id,
  regime,
  source
)
}
\arguments{
\item{loc_html}{A character vector which designates the location of the html files to extract, e.g. in a folder called "aze_speech/"}

\item{x_text}{The xpath of the text (to be found out, for example, by using the "inspect" tool of firefox)}

\item{x_title}{The xpath of the title of the text
(to be found out, for example, by using the "inspect" tool of firefox)}

\item{x_date}{The xpath of the text's date, if available
(to be found out, for example, by using the "inspect" tool of firefox).
One might like to use regular expressions and/or the as.Date function
to further adjust the format of the date once extracted.}

\item{speaker}{The name of the speaker (assuming official speeches are extracted from htmls)}

\item{country}{The name of the country of the speaker
(assuming official speeches are extracted from htmls)}

\item{country_id}{A short country identifier,
useful for merging the speech corpus with other datasets during the analysis.}

\item{regime}{The regime type of the speaker's country
(assuming official speeches are extracted from htmls), e.g. autocracy.}

\item{source}{The webpage from which the htmls were scraped
(crucial to always indicate due to copyright reasons!)}
}
\value{
A quanteda corpus consisting of all texts scraped as corpus documents
(for quanteda tutorials on text analysis see https://quanteda.io/index.html)
}
\description{
A function to extract texts, typically speeches,
scraped as html files and ideally stored in a folder in your working directory
(= the result of using the functions get_links, get_pagesource, and get_html).
}
\examples{
#Don't run
#Extract texts from htmls
#get_text()

}
