% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links_content.R
\name{get_links}
\alias{get_links}
\title{A function to \emph{automatically collect specific links on a website, e.g. to public speeches on official websites}.}
\usage{
get_links(website, start_website, page_ex, links_xpath, case, n)
}
\arguments{
\item{website}{The target website with the links to the desired content.}

\item{page_ex}{The page counter in the link of the main website which - most likely - refers to the links with content on several pages.
Typically an expression such as "page/30" at the end of the link.}

\item{links_xpath}{The general xpath node which contains the links of interest, often similar to "//h2//a".}

\item{case}{The name of the case, e.g. a country, make it a unique case identifier.}

\item{n}{The number of pages on the target website which contains the links of interest.}

\item{start_webpage}{The main webpage of the case of interest.}
}
\value{
A character vector with all links ("all_links)
}
\description{
A function to \emph{automatically collect specific links on a website, e.g. to public speeches on official websites}.
}
\examples{
#Don't run
#Get the links provided on a website
#get_links()

}
